requirements/base.txt
# Bleeding edge Django
django==1.8.7  # rq.filter: >=1.8,<1.9

# Configuration
django-environ==0.4.0
django-secure==1.0.1
whitenoise==2.0.6


# Forms
django-braces==1.8.1
django-crispy-forms==1.5.2
django-floppyforms==1.5.2

# Models
django-model-utils==2.3.1

# Images
Pillow==3.0.0

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.24.1


# Python-PostgreSQL Database Adapter
psycopg2==2.6.1

# Unicode slugification
unicode-slugify==0.1.3
django-autoslug==1.9.3

# Time zones support
pytz==2015.7

# Redis support
django-redis==4.3.0
redis>=2.10.0



# Your custom requirements go here

# django-cms
beautifulsoup4==4.4.0
cmsplugin-filer==0.10.2
django-admin-sortable2==0.6.1
django-angular==0.8.1
django-appconf==1.0.1
django-classy-tags==0.6.2
django-cms>=3.2.0
django-compressor==1.5
django-filer==1.0.6
django-filter==0.11.0
django-fsm==2.2.1
django-fsm-admin==1.2.2
django-haystack==2.4.0
django-ipware==1.1.1
django-libsass==0.4
django-mptt==0.7.4
django-parler==1.5
django-polymorphic==0.8.1
django-post-office==2.0.5
django-rest-auth==0.5.0
django-sass-processor==0.3.3
django-sekizai==0.8.2
django-select2==5.5.0
djangoshop-stripe==0.1.1
django-treebeard==3.0
djangocms-admin-style==0.2.8
git+https://github.com/jrief/djangocms-bootstrap3@fccfa3e1530ae2a151f58442de46526c82c1970b
djangocms-cascade==0.7.1
djangocms-text-ckeditor==2.6.0
djangorestframework==3.1.3
drf-haystack==1.5.2
easy-thumbnails==2.3
elasticsearch==1.7.0
html5lib==0.9999999
jsonfield==1.0.3
libsass==0.8.3
oauthlib==1.0.3
pluggy==0.3.1
py==1.4.30
Pygments==2.0.2
python-openid==2.2.5
requests==2.7.0
requests-oauthlib==0.5.0
six==1.9.0
stripe==1.27.1
Unidecode==0.4.18
urllib3==1.12
wheel==0.24.0
